scripts/requirements.txt
greenlet==2.0.2
numpy==1.24.2
pandas==2.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==1.4.16
typing-extensions==4.5.0
tzdata==2023.3
